+++ src/main/java/com/birca/bircabackend/command/auth/authentication/RequiredLoginInterceptor.java
@@ -1,4 +1,4 @@
-package com.birca.bircabackend.command.auth.login;
+package com.birca.bircabackend.command.auth.authentication;
 
 import com.birca.bircabackend.command.auth.exception.AuthErrorCode;
 import com.birca.bircabackend.command.auth.token.JwtTokenExtractor;
