--- src/main/java/com/birca/bircabackend/command/auth/login/LoginMember.java
@@ -1,4 +1,0 @@
-package com.birca.bircabackend.command.auth.login;
-
-public record LoginMember(Long id) {
-}
