+++ src/test/java/com/birca/bircabackend/command/member/application/MemberServiceTest.java
@@ -1,6 +1,6 @@
 package com.birca.bircabackend.command.member.application;
 
-import com.birca.bircabackend.command.auth.login.LoginMember;
+import com.birca.bircabackend.command.auth.authentication.LoginMember;
 import com.birca.bircabackend.command.member.MemberFixtureRepository;
 import com.birca.bircabackend.command.member.dto.NicknameRegisterRequest;
 import com.birca.bircabackend.command.member.dto.RoleChangeRequest;
